+++ metamon/data/replay_dataset/raw_replays/replay_intake.py
@@ -1,0 +1,463 @@
+import os
+import json
+import argparse
+import shutil
+from typing import Dict, List, Tuple, Set
+from collections import defaultdict
+from datetime import datetime
+import multiprocessing as mp
+from functools import partial
+import re
+
+
+def find_json_files(directory: str) -> List[str]:
+    json_files = []
+    for root, _, files in os.walk(directory):
+        for file in files:
+            if file.endswith(".json"):
+                json_files.append(os.path.join(root, file))
+    return json_files
+
+
+def extract_game_id(json_data: dict) -> str:
+    return str(json_data.get("id", ""))
+
+
+def normalize_format(format_str: str) -> str:
+    format_str = format_str.lower()
+    gen_num = None
+    gen_bracket_match = re.search(r"\[gen\s*(\d+)\]", format_str)
+    if gen_bracket_match:
+        gen_num = gen_bracket_match.group(1)
+    if not gen_num:
+        gen_direct_match = re.search(r"gen(\d+)", format_str)
+        if gen_direct_match:
+            gen_num = gen_direct_match.group(1)
+    if not gen_num:
+        return format_str
+    tier_match = re.search(r"(ou|uu|nu|ubers)", format_str)
+    tier = tier_match.group(1) if tier_match else None
+    if not tier:
+        return format_str
+    return f"gen{gen_num}{tier}"
+
+
+def get_existing_game_ids(target_dir: str) -> Dict[str, int]:
+    """
+    Get dictionary mapping game IDs to upload times for replays that already exist in target directory.
+    Returns Dict[game_id, upload_time].
+    """
+    existing_games = {}
+    if not os.path.exists(target_dir):
+        return existing_games
+
+    for gen_dir in os.listdir(target_dir):
+        gen_path = os.path.join(target_dir, gen_dir)
+        if not os.path.isdir(gen_path):
+            continue
+        for tier_dir in os.listdir(gen_path):
+            tier_path = os.path.join(gen_path, tier_dir)
+            if not os.path.isdir(tier_path):
+                continue
+            for json_file in os.listdir(tier_path):
+                if json_file.endswith(".json"):
+                    json_path = os.path.join(tier_path, json_file)
+                    try:
+                        with open(json_path, "r") as f:
+                            data = json.load(f)
+                            game_id = extract_game_id(data)
+                            upload_time = int(data.get("uploadtime", 0))
+                            if game_id:  # Only store if we have a valid game ID
+                                existing_games[game_id] = upload_time
+                    except (json.JSONDecodeError, IOError):
+                        continue
+    return existing_games
+
+
+def get_format_dirs(format_str: str) -> Tuple[str, str]:
+    """Extract generation and tier directories from format string."""
+    gen_match = re.search(r"gen(\d+)", format_str)
+    if not gen_match:
+        return None, None
+
+    gen_dir = f"gen{gen_match.group(1)}"
+    tier = format_str.replace(gen_dir, "").strip()
+    return gen_dir, tier
+
+
+def copy_to_target(
+    source_path: str,
+    game_id: str,
+    format_str: str,
+    upload_time: int,
+    target_base: str,
+    existing_games: Dict[str, int],
+) -> bool:
+    """
+    Copy a JSON file to the appropriate target directory.
+    Returns True if file was copied, False if skipped (already exists).
+    """
+    if game_id in existing_games:
+        return False
+
+    gen_dir, tier = get_format_dirs(format_str)
+    if not gen_dir or not tier:
+        return False
+
+    # Create directory structure
+    tier_path = os.path.join(target_base, gen_dir, tier)
+    os.makedirs(tier_path, exist_ok=True)
+
+    # Define target file path using game ID
+    target_path = os.path.join(tier_path, f"{game_id}.json")
+
+    # Copy the file
+    shutil.copy2(source_path, target_path)
+    return True
+
+
+def process_single_file(json_path: str, allowed_formats: set) -> Dict:
+    """Process a single JSON file and return its statistics if valid."""
+    try:
+        with open(json_path, "r") as f:
+            json_data = json.load(f)
+
+        # Basic validation checks
+        if not json_data.get("log"):
+            return {"valid": False}
+
+        # Check for game ID
+        game_id = str(json_data.get("id", ""))
+        if not game_id:
+            return {"valid": False}
+
+        log_lines = json_data["log"].split("\n")
+
+        # Check for required log content
+        has_win = any(line.startswith("|win|") for line in log_lines)
+        has_gen = any(line.startswith("|gen|") for line in log_lines)
+        has_tier = any(line.startswith("|tier|") for line in log_lines)
+
+        # Check for minimum number of turns (to filter out forfeits/disconnects)
+        turn_count = sum(1 for line in log_lines if line.startswith("|turn|"))
+        has_min_turns = turn_count >= 5
+
+        # Check for required metadata
+        has_format = "format" in json_data
+        has_upload_time = "uploadtime" in json_data and int(json_data["uploadtime"]) > 0
+
+        # Validate format
+        raw_format = json_data.get("format", "unknown")
+        normalized_format = normalize_format(raw_format)
+
+        # Skip if format doesn't match filter
+        if allowed_formats and normalized_format not in allowed_formats:
+            return {"valid": False}
+
+        # Combine all validation checks
+        is_valid = all(
+            [
+                has_win,
+                has_gen,
+                has_tier,
+                has_min_turns,
+                has_format,
+                has_upload_time,
+            ]
+        )
+
+        if not is_valid:
+            return {"valid": False}
+
+        # Only collect stats for valid replays
+        return {
+            "valid": True,
+            "normalized_format": normalized_format,
+            "upload_time": int(json_data["uploadtime"]),
+            "game_id": game_id,
+            "source_path": json_path,
+        }
+
+    except Exception as e:
+        return {"valid": False}
+
+
+def combine_stats(results: List[Dict]) -> Dict:
+    """Combine statistics from valid replays only."""
+    combined = {
+        "total_processed": len(results),
+        "valid_replays": 0,
+        "invalid_replays": 0,
+        "normalized_formats": defaultdict(int),
+        "unique_games": set(),  # Set of unique game IDs
+        "earliest_replay": float("inf"),
+        "latest_replay": 0,
+        "replays_by_year": defaultdict(int),
+    }
+
+    for stats in results:
+        if not stats["valid"]:
+            combined["invalid_replays"] += 1
+            continue
+
+        combined["valid_replays"] += 1
+        combined["normalized_formats"][stats["normalized_format"]] += 1
+
+        game_id = stats["game_id"]
+        upload_time = stats["upload_time"]
+
+        # Track unique games by ID only
+        combined["unique_games"].add(game_id)
+
+        if upload_time > 0:
+            combined["earliest_replay"] = min(combined["earliest_replay"], upload_time)
+            combined["latest_replay"] = max(combined["latest_replay"], upload_time)
+            year = datetime.fromtimestamp(upload_time).year
+            combined["replays_by_year"][year] += 1
+
+    return combined
+
+
+def generate_time_series(results: List[Dict]) -> Dict[str, List[Tuple[str, int]]]:
+    """
+    Generate monthly time series data for each format.
+    Returns Dict[format, List[Tuple[month_str, count]]] where month_str is YYYY-MM.
+    """
+    # Track monthly counts for each format
+    format_monthly_counts = defaultdict(lambda: defaultdict(int))
+
+    # Process each valid replay
+    for result in results:
+        if not result["valid"]:
+            continue
+
+        format_str = result["normalized_format"]
+        timestamp = result["upload_time"]
+        if timestamp > 0:
+            # Convert timestamp to YYYY-MM format
+            month_str = datetime.fromtimestamp(timestamp).strftime("%Y-%m")
+            format_monthly_counts[format_str][month_str] += 1
+
+    # Convert to sorted lists of (month, count) tuples
+    time_series = {}
+    for format_str, monthly_counts in format_monthly_counts.items():
+        # Sort by month string (YYYY-MM format ensures chronological order)
+        sorted_months = sorted(monthly_counts.items())
+        time_series[format_str] = sorted_months
+
+    return time_series
+
+
+def save_time_series(
+    time_series: Dict[str, List[Tuple[str, int]]], output_path: str
+) -> None:
+    """Save time series data to a JSON file."""
+    # Convert tuples to lists for JSON serialization
+    json_data = {
+        format_str: [{"month": month, "count": count} for month, count in monthly_data]
+        for format_str, monthly_data in time_series.items()
+    }
+
+    # Add metadata
+    json_data["_metadata"] = {
+        "created_at": datetime.now().isoformat(),
+        "total_formats": len(json_data) - 1,  # Subtract 1 for metadata
+        "date_range": {
+            "start": min(
+                data[0]["month"]
+                for data in json_data.values()
+                if isinstance(data, list) and data
+            ),
+            "end": max(
+                data[-1]["month"]
+                for data in json_data.values()
+                if isinstance(data, list) and data
+            ),
+        },
+    }
+    # Save to file
+    with open(output_path, "w") as f:
+        json.dump(json_data, f, indent=2)
+
+
+def analyze_replays(
+    directory: str,
+    allowed_formats: set,
+    target_dir: str = None,
+    num_processes: int = None,
+    time_series_path: str = None,
+) -> None:
+    """Analyze all replay JSONs in a directory using parallel processing."""
+    json_files = find_json_files(directory)
+    total_files = len(json_files)
+    print(f"\nFound {total_files} JSON files")
+
+    if total_files == 0:
+        print("No JSON files found in the specified directory")
+        return
+
+    # Get existing game IDs and upload times if target directory is specified
+    existing_games = get_existing_game_ids(target_dir) if target_dir else {}
+    if target_dir:
+        print(f"Found {len(existing_games)} existing replays in target directory")
+
+    # Determine number of processes
+    if num_processes is None:
+        # Default to all available CPU cores except one
+        num_processes = max(1, mp.cpu_count() - 1)
+    else:
+        # Ensure num_processes is within valid range
+        num_processes = max(1, min(num_processes, mp.cpu_count()))
+
+    print(f"Processing files using {num_processes} processes...")
+
+    # Process files in parallel
+    with mp.Pool(num_processes) as pool:
+        # Create a partial function with the allowed_formats
+        process_func = partial(process_single_file, allowed_formats=allowed_formats)
+        # Use imap_unordered for better performance with large datasets
+        results = list(
+            pool.imap_unordered(
+                process_func,
+                json_files,
+                chunksize=max(1, total_files // (num_processes * 10)),
+            )
+        )
+
+    # Generate and save time series if requested
+    if time_series_path:
+        time_series = generate_time_series(results)
+        save_time_series(time_series, time_series_path)
+        print(f"\nSaved time series data to: {time_series_path}")
+
+    # Combine results from all processes
+    stats = combine_stats(results)
+
+    # Copy valid files if target directory specified
+    if target_dir:
+        # Deduplicate results by game ID, keeping most recent version
+        deduplicated_results = {}
+        for result in results:
+            if not result["valid"]:
+                continue
+            game_id = result["game_id"]
+            upload_time = result["upload_time"]
+
+            # Keep this result if it's the first time we've seen this game_id
+            # or if it's more recent than the previous one we've seen
+            if (
+                game_id not in deduplicated_results
+                or upload_time > deduplicated_results[game_id]["upload_time"]
+            ):
+                deduplicated_results[game_id] = result
+
+        print(
+            f"\nDeduplicated {len(results)} replays to {len(deduplicated_results)} unique games"
+        )
+
+        copied_count = 0
+        skipped_count = 0
+        updated_count = 0
+
+        for result in deduplicated_results.values():
+            game_id = result["game_id"]
+            upload_time = result["upload_time"]
+
+            # Skip if game already exists in target
+            if game_id in existing_games:
+                skipped_count += 1
+                continue
+
+            if copy_to_target(
+                result["source_path"],
+                game_id,
+                result["normalized_format"],
+                upload_time,
+                target_dir,
+                existing_games,
+            ):
+                copied_count += 1
+
+    # Print summary
+    print("\nAnalysis Summary:")
+    if allowed_formats:
+        print(f"Filtering for formats: {', '.join(sorted(allowed_formats))}")
+    print(f"Total files processed: {stats['total_processed']}")
+    print(f"Valid replays: {stats['valid_replays']}")
+    print(f"Invalid replays: {stats['invalid_replays']}")
+    print(f"Unique games (by ID only): {len(stats['unique_games'])}")
+
+    if target_dir:
+        print(f"\nFile Operations:")
+        print(f"  Copied to target: {copied_count}")
+        print(f"  Skipped (already exists): {skipped_count}")
+
+    if stats["earliest_replay"] != float("inf"):
+        print("\nTimestamp Range (valid replays only):")
+        earliest = datetime.fromtimestamp(stats["earliest_replay"])
+        latest = datetime.fromtimestamp(stats["latest_replay"])
+        print(f"  Earliest: {earliest.strftime('%Y-%m-%d %H:%M:%S')}")
+        print(f"  Latest: {latest.strftime('%Y-%m-%d %H:%M:%S')}")
+
+        print("\nReplays by Year:")
+        for year in sorted(stats["replays_by_year"].keys()):
+            print(f"  {year}: {stats['replays_by_year'][year]}")
+
+    print("\nFormat distribution:")
+    sorted_formats = sorted(
+        stats["normalized_formats"].items(), key=lambda x: x[1], reverse=True
+    )
+    for format_id, count in sorted_formats:
+        print(f"  {format_id}: {count}")
+
+
+def main():
+    parser = argparse.ArgumentParser(
+        description="Analyze Pokemon Showdown replay JSON files"
+    )
+    parser.add_argument(
+        "directory", type=str, help="Directory containing replay JSON files"
+    )
+    parser.add_argument(
+        "--formats",
+        type=str,
+        nargs="*",
+        default=[
+            f"gen{i}{f}" for i in range(1, 5) for f in ["ou", "uu", "nu", "ubers"]
+        ],
+        help="Only process replays matching these formats (e.g., gen1ou gen1ubers). Default: process all formats",
+    )
+    parser.add_argument(
+        "--target-dir",
+        type=str,
+        help="Target directory to copy valid replays into an organized structure",
+    )
+    parser.add_argument(
+        "--processes",
+        type=int,
+        help="Number of processes to use for parallel processing. Default: number of CPU cores minus 1",
+    )
+    parser.add_argument(
+        "--save-time-series",
+        type=str,
+        help="Path to save monthly time series data for each format (as JSON)",
+    )
+    args = parser.parse_args()
+
+    if not os.path.isdir(args.directory):
+        print(f"Error: {args.directory} is not a valid directory")
+        return
+
+    # Convert formats list to set for faster lookup
+    allowed_formats = set(args.formats) if args.formats else set()
+    analyze_replays(
+        args.directory,
+        allowed_formats,
+        args.target_dir,
+        args.processes,
+        args.save_time_series,
+    )
+
+
+if __name__ == "__main__":
+    main()
